Do not add extra_params to the lifted_cont_params (#3885)
Currently, any and all variables in scope, including parameters
and extra parameters (coming from CSE and unboxing), are added to the
lifted_cont_params. This is superfluous, since the handlers of lifted
continuations do not contain any reference to these extra params prior
to simplification, and these extra params can be recomputed for the
lifted continuations.

As it stands, this commit alone may lead to regressions, since we
currently do not perform unboxing on the lifted_cont_params. This will
be fixed in a later commit.