Avoid needless dependency on parameter .cmi by introducing `Parameter_name.t` (#3708)
* Introduce `Global_module.Parameter_name.t`

Takes the place of any `Global_module.t` or `Global_module.Name.t` that is known
to be the name of a parameter. Can't have parameters. (Notably, even the current
half-worked-out plans for parameterized parameters would actually keep this type
in many places, so this isn't as much in the way as one might think.)

* Don't call `global_of_global_name` on parameter names

This has the benefit of not loading the .cmi file for something whose signature
we don't need (say because the parameter is only used as part of an instance
name).

* Code review