Prevent continuation specialization when there is a soc in the handler (#4751)
* Add test

* Prevent continuation specialization whzen there is a soc in the handler

When there is a set of closures in the handler of a continuation, as
well as an inlinable call to that set of closures, specialization of
that handler can result in lifting a continuation that has already eben
simplified in the context of the handler, and can thus contain
references to lifted constants (which can be either symbols, or code
ids). In such a situation, the lifted continuation makes these lifted
constants "escape" from the first downward traversal, which can result
in errors later, such as a missing code or missing symbol fatal error.

For now we simply disable specialization of continuations that contain
sets of closures. In the future we may want to try and properly manage
to extract the lifted constants from the first downward pass, which
would make it possible to specialize continuation whose handler contain
a set of closures.